update real experiment

+++ warpfunc/indicator.m
@@ -1,4 +1,4 @@
 function z=indicator(c,x)
     % return I{x>c}
-    z=x>c;
+    z= double(x>c);
 end
